# swagger/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: SaltRiver
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
  - multipart/form-data
  - application/x-www-form-urlencoded
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /note:
    x-swagger-router-controller: note
    get:
      tags:
        - note
      description: 전체 노트 받아오기
      operationId: getNotes
      security:
        - jwt: []
      parameters:
        - name: note_idx
          in: query
          description: 조회하고 싶은 특정 노트 인덱스
          required: false
          type: string
        - name: url
          in: query
          description: 사용자가 조회하고 싶은 url 정보
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/RelationResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      tags:
        - note
      description: 신규 노트 등록시 작동하는 api
      operationId: addNote
      security:
        - jwt: []
      parameters:
        - name: info
          in: body
          description: Note Info
          schema:
            type: object
            properties:
              url:
                type: string
                description: 노트한 사이트 url
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/RelationResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    patch:
      tags:
        - note
      description: 노트 업데이트
      operationId: patchNote
      summary: 노트 업데이트
      security:
        - jwt: []
      parameters:
        - name: info
          in: body
          description: Note Info
          schema:
            type: object
            properties:
              idx:
                type: integer
                description: 상태를 변경하고자 하는 관계의 인덱스
              url:
                type: string
                description: 노트한 사이트 url
              draw_url:
                type: string
                description: draw 파일
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/RelationResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      tags:
        - note
      description: Delete relation API
      operationId: deleteNote
      # security:
      #   - jwt: []
      parameters:
        - name: idx
          in: query
          description: Index that we wanted to delete
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/RelationResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /note/share:
    x-swagger-router-controller: note
    get:
      tags:
        - note
      description: 전체 노트 받아오기
      operationId: getShareNotes
      security:
        - jwt: []
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/RelationResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      tags:
        - note
      description: 공유 리스트
      operationId: addShare
      security:
        - jwt: []
      parameters:
        - name: info
          in: body
          description: Note Info
          schema:
            type: object
            properties:
              id:
                type: string
                description: 공유 하고자 하는 상대의 아이디값
              note_idx:
                type: integer
                description: 노트 고유 idx 값
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/RelationResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      tags:
        - note
      description: Delete relation API
      operationId: disconnectSharing
      security:
        - jwt: []
      parameters:
        - name: idx
          in: query
          description: Index that we wanted to delete
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/RelationResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /note/shared:
    x-swagger-router-controller: note
    get:
      tags:
        - note
      description: 전체 노트 받아오기
      operationId: getSharedNoteContents
      security:
        - jwt: []
      parameters:
        - name: url
          in: query
          description: 사용자가 방문한 url 정보
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/RelationResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /note/draw:
    x-swagger-router-controller: note
    post:
      tags:
        - note
      description: 필기 내용을 입력할 때 사용되는 url
      operationId: addDraw
      # security:
      #   - jwt: []
      parameters:
        - name: info
          in: body
          description: Note Info
          schema:
            type: object
            properties:
              note_idx:
                type: integer
                description: 사용자가 메모를 한 노트의 인덱스
              draw_url:
                type: string
                description: 필기를 해서 생성된 파일의 url
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/RelationResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /note/memo:
    x-swagger-router-controller: note
    post:
      tags:
        - note
      description: 메모를 새로 추가할 때 사용되는 api로 return 값으로 memo 고유의 인덱스(idx)값을 준다. 해당 메모에 대한 수정등의 요청시 해당 인덱스 값을 보내주어야 한다.
      operationId: addMemo
      # security:
      #   - jwt: []
      parameters:
        - name: info
          in: body
          description: Note Info
          schema:
            type: object
            properties:
              note_idx:
                type: integer
                description: 사용자가 메모를 한 노트의 인덱스
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/RelationResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    patch:
      tags:
        - note
      description: 기존에 사용된 메모 내용을 수정할 때 사용되는 api
      operationId: patchMemo
      # security:
      #   - jwt: []
      parameters:
        - name: info
          in: body
          description: Note Info
          schema:
            type: object
            properties:
              memo_idx:
                type: integer
                description: 사용자가 메모를 한 노트의 인덱스
              content:
                type: string
                description: 사용자 메모 내용
              top:
                type: number
                description: 메모의 좌표 값(y)
              left:
                type: number
                description: 메모의 좌표 값(x)
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/RelationResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      tags:
        - note
      description: 기존에 사용된 메모 내용을 수정할 때 사용되는 api
      operationId: deleteMemo
      # security:
      #   - jwt: []
      parameters:
        - name: memo_idx
          in: query
          description: 메모 인덱스
          required: true
          type: integer
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/RelationResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
securityDefinitions:
  jwt:
    type: apiKey
    in: header
    name: jwt
# complex objects have schema definitions
definitions:
...
  RelationResponse:
    properties:
      data:
        type: string
...
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string